refactor(package-manager): give each module one install phase
The first split left the decomposition arbitrary: `linking` was a module
while the resolve pass stayed inline, and two functions sat in modules
that did not own their phase. A reader could not infer any rule for what
became a module, which is worse than either extreme.

Make the rule "one module per install phase" and hold to it:

- `resolution_inputs` was a grab-bag — the read-package transform chain,
  the preferred-versions seeds, the reuse seed, the shared resolve
  options, and a post-resolve diagnostic. Split into
  `manifest_transforms` (the `packageExtensions` + `pnpm.overrides`
  chain) and `resolve` (everything the walk consumes, the walk, and the
  diagnostics that read its result).
- `warn_stale_convergence_overrides` reads the resolve's output, so it
  belongs in `resolve`, not among its inputs.
- The resolve pass itself moves into `resolve::run_resolve_pass`, so the
  phase is a module like the others rather than 130 inline lines.
- `build_extra_env` is build-phase setup, not linking. It moves beside
  its only caller. The build phase deliberately does *not* become a
  module here: `install_frozen_lockfile::run_build_phase` already owns
  it and is shared with the frozen path, so a local wrapper would add a
  layer and no meaning.

`run` is 921 lines, down from 2103.
